refactor: phase 0 - detach cache layers
