+++ app/lib/cache/index.ts
@@ -697,12 +697,3 @@ export async function fetchCached<T>(
   inflightMap.set(cacheKey, entry);
   return request;
 }
-
-export function isUniqueConstraintError(err: Error): { table: string; column: string } | null {
-  const match = err.message.match(/UNIQUE constraint failed: (\w+)\.(\w+)/);
-  if (match) {
-    return { table: match[1], column: match[2] };
-  }
-
-  return null;
-}
