+++ app/lib/db.ts
@@ -1,0 +1,8 @@
+export function isUniqueConstraintError(err: Error): { table: string; column: string } | null {
+  const match = err.message.match(/UNIQUE constraint failed: (\w+)\.(\w+)/);
+  if (match) {
+    return { table: match[1], column: match[2] };
+  }
+
+  return null;
+}
