+++ docs/architecture.md
@@ -3,6 +3,10 @@
 몰루로그의 문서는 파일 목록보다 오래 유지되는 구조와 규칙을 설명하는 것을 목표로 합니다.
 이 문서도 현재 코드의 세부 나열보다 "어디에 무엇을 두는가"와 "데이터가 어떻게 흐르는가"에 집중합니다.
 
+> 이 문서는 목표 아키텍처(최종 형태)를 기준으로 작성합니다.
+> 데이터 접근 계층은 현재 단계적으로 이관하는 중이라 일부 코드는 아직 과도기 형태로 남아 있습니다.
+> 새 코드는 과도기 코드가 아니라 이 문서의 목표 구조를 기준으로 작성합니다.
+
 ## 전체 구조
 
 ```text
@@ -21,18 +25,32 @@ Cloudflare D1 / KV
 
 ## 핵심 디렉터리
 
+데이터는 위에서 아래로 한 방향으로만 흐릅니다.
+
+```text
+Routes  (loader/action · 얇게)
+  → Views   (합성 + 라우트 캐시 SWR)
+      → Domain  (순수 계산 · I/O 없음)
+      → Models  (D1 CRUD · BAQL 읽기 · 소스 캐시 · 정규화)
+          → lib/cache · lib/baql · db
+```
+
 - `app/routes`
-  라우트 파일, `loader`/`action`, 메타, 상위 화면 조립을 담당합니다.
+  라우트 파일과 `loader`/`action`, 메타, 파라미터 해석, 인증, 화면 조립을 맡습니다. 데이터 합성이나 캐시·BAQL·D1 직접 호출은 라우트에 두지 않고 뷰에 맡깁니다.
+- `app/views`
+  라우트 표현용 합성 레이어입니다. 여러 모델·도메인 결과를 화면이 필요한 모양으로 합치고 라우트 캐시(SWR)를 적용합니다.
+- `app/models`
+  데이터 접근 전용 레이어입니다. D1 CRUD, BAQL 읽기, 소스 캐시, upstream 데이터의 도메인 타입 정규화까지 담당합니다.
+- `app/domain`
+  순수 계산과 변환만 담는 레이어입니다. I/O와 env 의존이 없어 단위 테스트하기 쉽습니다. 레이드 점수 계산이나 모집 시뮬레이터 같은 로직이 여기에 옵니다.
+- `app/lib/cache`, `app/lib/baql`, `app/lib/db`
+  캐시 원시 함수, GraphQL 실행, D1 헬퍼 같은 인프라입니다.
 - `app/components/primitives`
   semantic token을 사용하는 저수준 공통 UI와 얇은 앱 공통 표현 레이어입니다.
 - `app/components/features/<domain>`
   여러 화면에서 재사용되는 도메인 UI입니다.
 - `app/routes/*._components`, `app/routes/*/_components`
   한 라우트 패밀리 안에서만 쓰는 route-local UI와 훅입니다.
-- `app/models`
-  D1 테이블 정의와 앱 도메인 단위의 읽기/쓰기 로직이 위치합니다.
-- `app/repositories`
-  새로 추가되는 BAQL 읽기 조합, 캐싱, 복합 조회 orchestration은 우선 이 계층을 고려합니다.
 - `workers/app.ts`
   Worker 엔트리와 크론 작업 진입점입니다.
 
@@ -67,9 +85,9 @@ Cloudflare D1 / KV
 ### 읽기
 
 1. 브라우저가 라우트를 요청합니다.
-2. route `loader`가 실행됩니다.
-3. 필요하면 KV 캐시를 먼저 확인합니다.
-4. 캐시 미스 시 BAQL 또는 D1을 조회합니다.
+2. route `loader`가 실행되어 해당 뷰 함수 하나를 호출합니다.
+3. 뷰는 필요한 모델·도메인을 호출해 화면 모양으로 합성하고 필요하면 라우트 캐시를 둡니다.
+4. 모델은 소스 캐시를 먼저 확인하고 미스면 BAQL 또는 D1을 조회합니다.
 5. `useLoaderData()`로 화면을 렌더링합니다.
 
 ### 쓰기
@@ -82,10 +100,13 @@ Cloudflare D1 / KV
 
 ## 캐싱 원칙
 
-- 공통 캐시 유틸리티는 `app/models/base.ts`의 `fetchCached`를 사용합니다.
+- 캐시 원시 함수는 `app/lib/cache`에 모읍니다. (`fetchCached`, `fetchSourceCached`, `fetchRouteCached` 등)
+- 캐시는 두 종류로 구분합니다.
+  - 소스 캐시(`fetchSourceCached`): 원본 BAQL 참조 데이터. 모델에서만 적용합니다.
+  - 라우트 캐시(`fetchRouteCached`, SWR): 합성한 화면 view-model. 뷰에서만 적용합니다.
 - 캐시 키는 `cache::` 접두사 아래에 저장됩니다.
-- 화면 단위에서 임의로 캐시 정책을 만들기보다, 재사용되는 조회 로직 안에서 캐시를 결정합니다.
-- 강제 새로고침이 필요한 경우 `forceRefresh` 패턴을 사용합니다.
+- 라우트나 화면에서 임의로 캐시 정책을 만들지 않고 책임 있는 레이어(모델/뷰) 안에서 결정합니다.
+- 강제 새로고침이 필요하면 `forceRefresh` 패턴을 사용합니다.
 
 ## 인증 원칙
